Batch Update
Update
1.	User
a.	Merapihkan Response createUser
b.	Update user memakai locals user (Questionable)
c.	Delete user memakai locals user (Questionable)
d.	Topup user memakai locals user Questionable)
e.	Menambahkan asosiasi (hasMany TransactionHistory)
2.	Category
a.	Penggunaan Autorisasi admin untuk mengakses API kategori
b.	Catch Error createCategory (Kategori sudah ada) menggunakan SequelizeUniqueConstraintError
c.	getCategory include model Product
d.	Catch Error updateCategory ketika ID Category pada Params tidak ada di database
e.	Catch Error deleteCategory ketika ID Category pada Params tidak ada di database
f.	Migrate tipe category memakai unique (untuk catch error kategori sudah ada)
3.	Product
a.	Penggunaan Autorisasi admin untuk mengakses API produk
b.	Response price product memakai format rupiah (Ex : Rp.10000)
c.	Catch error ketika kategori yang dimasukkan ke produk tidak terdaftar menggunakan SequelizeForeignKeyConstraintError
d.	Catch Error updateProduct ketika ID Product pada Params tidak ada di database
e.	Catch Error deleteProduct ketika ID Product pada Params tidak ada di database
4.	TransactionHistory
a.	Cek produk menggunakan id
b.	Cek Stok barang
c.	Cek saldo user mencukupi
d.	Update saldo user, stok produk, dan sold amount produk
e.	getUserTransaction sudah bisa
f.	getAdminTransaction menggunakan Autorisasi admin untuk akses
g.	getTransaction

+++ Controllers/CatagoryController.js
@@ -1,29 +1,20 @@
-const {Category} = require('../models');
+const {Category,Product} = require('../models');
 
 class CategoryController{
     static async createCategory (req,res){
-        const {tipe, sold_product_amount} = req.body;
         try {
-            const cekTipe = await Category.findOne({
-                where: {tipe}
-            });
+            const create = await Category.create(req.body);
 
-            if(!cekTipe){
-                const create = await Category.create({
-                    tipe, 
-                    sold_product_amount
-                });
-
-                res.status(201).json({
-                    message: "Category Berhasil di tampilkan",
-                    data: create
-                })
-            }else{
-                res.status(400).json({
-                    message: "Tipe tersebut sudah ada"
+            res.status(201).json({
+                category: create
+            })
+        } catch (error) {
+            const { name } = error
+            if(name === "SequelizeUniqueConstraintError"){
+                return res.status(400).json({
+                    message: error.message
                 })
             }
-        } catch (error) {
             res.status(404).json({
                 message: error.message
             })
@@ -32,15 +23,18 @@ class CategoryController{
 
     static async getCategory (req,res){
         try {
-            const get = await Category.findAll();
+            const get = await Category.findAll({
+                include: [{
+                    model: Product
+                }]
+            });
             if(get.length>0){
                 res.status(200).json({
-                    message: "Menampilkan Data Category",
-                    data: get
+                    categories: get
                 })
             }else{
                 res.status(200).json({
-                    message: "Belum ada Category"
+                    message: "Category Not Found!"
                 })
             }
         } catch (error) {
@@ -51,20 +45,26 @@ class CategoryController{
     }
 
     static async updateCategory (req,res){
-        const {tipe, sold_product_amount} = req.body;
         try {
-            await Category.update({
-                tipe, 
-                sold_product_amount
-            }, {
-                where: {
-                    id: req.params.id
-                }
-            });
-
-            res.status(200).json({
-                message: "Data berhasiil di update"
+            const find = await Category.findOne({
+                attributes: ['id'],
+                where: {id: req.params.id}
             })
+            if(find){
+                await Category.update(req.body, {
+                    where: {id: req.params.id}
+                });
+                const getUpdate = await Category.findOne({
+                    where: {id: req.params.id}
+                })
+                res.status(200).json({
+                    category: getUpdate
+                })
+            } else {
+                res.status(400).json({
+                    message: "There is No Such Category With ID "+req.params.id
+                })
+            }
         } catch (error) {
             res.status(404).json({
                 message: error.message
@@ -74,14 +74,22 @@ class CategoryController{
 
     static async deleteCategory (req,res){
         try {
-            await Category.destroy({
-                where: {
-                    id: req.params.id
-                }
+            const find = await Category.findOne({
+                attributes: ['id'],
+                where: {id: req.params.id}
             })
-            res.status(200).json({
-                message: "Data Berhasil di Hapus"
-            })
+            if(find){
+                await Category.destroy({
+                    where: {id: req.params.id}
+                })
+                res.status(200).json({
+                    message: "Category Has Been Successfully Deleted"
+                })
+            } else {
+                res.status(400).json({
+                    message: "There is No Such Category With ID "+req.params.id
+                })
+            }
         } catch (error) {
             res.status(404).json({
                 message: error.message
