+++ Controllers/ProductController.js
@@ -2,20 +2,19 @@ const {Product} = require('../models');
 
 class ProductController{
     static async createProduct (req,res){
-        const {title,price,stock, CatagoryId} = req.body;
         try {
-            const product = await Product.create({
-                title,
-                price,
-                stock, 
-                CatagoryId
-            });
-
+            const product = await Product.create(req.body);
+            product.price = "Rp."+product.price
             res.status(201).json({
-                message: "Data Berhasil di tambahkan",
-                data: product
+                product: product
             })
         } catch (error) {
+            const { name } = error
+            if(name === "SequelizeForeignKeyConstraintError"){
+                return res.status(400).json({
+                    message: "There is No Such Category With ID "+ req.body.CategoryId
+                })
+            }
             res.status(404).json({
                 message: error.message
             })
@@ -26,13 +25,15 @@ class ProductController{
         try {
             const get = await Product.findAll();
             if(get.length>0){
+                for(const i in get){
+                    get[i].price = "Rp."+get[i].price
+                }
                 res.status(200).json({
-                    message: "Menampilkan Product",
                     Product: get
                 })
             }else{
                 res.status(200).json({
-                    message: "Tidak ada Product"
+                    message: "No products listed!"
                 })
             }
         } catch (error) {
@@ -44,24 +45,34 @@ class ProductController{
 
 
     static async updateProduct (req,res){
-        const {title,price,stock, CatagoryId} = req.body;
         try {
-            await Product.update({
-                title,
-                price,
-                stock, 
-                CatagoryId
-            },{
-                where: {
-                    id: req.params.id
-                }
-            });
-
-            res.status(200).json({
-                message: "Product sudah di edit"
+            const find = await Product.findOne({
+                attributes: ['id'],
+                where: {id: req.params.id}
             })
-            
+            if(find){
+                await Product.update(req.body,{
+                    where: {id: req.params.id}
+                });
+                const getUpdate = await Product.findOne({
+                    where:{id: req.params.id}
+                })
+                getUpdate.price = "Rp."+getUpdate.price
+                res.status(200).json({
+                    product: getUpdate
+                })
+            } else {
+                res.status(400).json({
+                    message: "There is No Such Product With ID "+req.params.id
+                })
+            }
         } catch (error) {
+            const { name } = error
+            if(name === "SequelizeForeignKeyConstraintError"){
+                return res.status(400).json({
+                    message: "There is No Such Category With ID "+ req.body.CategoryId
+                })
+            }
             res.status(404).json({
                 message: error.message
             })
@@ -70,16 +81,23 @@ class ProductController{
 
     static async deleteProduct (req,res){
         try {
-            await Product.destroy({
-                where: {
-                    id: req.params.id
-                }
-            });
-
-            res.status(200).json({
-                message: "Data berhasil di hapus"
+            const find = await Product.findOne({
+                attributes: ['id'],
+                where: {id: req.params.id}
             })
-            
+            if(find){
+                await Product.destroy({
+                    where: {id: req.params.id}
+                });
+    
+                res.status(200).json({
+                    message: "Product Has Been Successfully Deleted"
+                })
+            } else {
+                res.status(400).json({
+                    message: "There is No Such Product With ID "+req.params.id
+                })
+            }
         } catch (error) {
             res.status(404).json({
                 message: error.message
